🚚 separate trait and its impl
