+++ lib/jobapi/src/lib.rs
@@ -2,3 +2,4 @@ pub mod actor;
 pub mod aws;
 pub mod grpc;
 pub mod jobapi;
+pub mod model;
